+++ src/Components/Persistent/BombDamage.hpp
@@ -2,13 +2,14 @@
 #define __CMP_BOMBDAMAGE_HPP__
 
 #include <Components/Persistent/BasePersistent.hpp>
-namespace ProceduralMaze::Cmp::Persistent {
+namespace ProceduralMaze::Cmp::Persistent
+{
 
-class BombDamage : public BasePersistent<int>
+class BombDamage : public BasePersistent<uint8_t>
 {
 public:
-  BombDamage( int value = 10, int min_value = 1, int max_value = 50 )
-      : BasePersistent<int>( value, min_value, max_value )
+  BombDamage( uint8_t value = 10, uint8_t min_value = 1, uint8_t max_value = 50 )
+      : BasePersistent<uint8_t>( value, min_value, max_value )
   {
   }
   virtual std::string class_name() const override { return "BombDamage"; }
