+++ src/Components/Persistent/CorruptionDamage.hpp
@@ -2,13 +2,14 @@
 #define __CMP_CORRUPTIONDAMAGE_HPP__
 
 #include <Components/Persistent/BasePersistent.hpp>
-namespace ProceduralMaze::Cmp::Persistent {
+namespace ProceduralMaze::Cmp::Persistent
+{
 
-class CorruptionDamage : public BasePersistent<int>
+class CorruptionDamage : public BasePersistent<uint8_t>
 {
 public:
-  CorruptionDamage( int value = 1, int min_value = 0, int max_value = 10 )
-      : BasePersistent<int>( value, min_value, max_value )
+  CorruptionDamage( uint8_t value = 1, uint8_t min_value = 0, uint8_t max_value = 10 )
+      : BasePersistent<uint8_t>( value, min_value, max_value )
   {
   }
   virtual std::string class_name() const override { return "CorruptionDamage"; }
