+++ src/Components/Persistent/GraveNumMultiplier.hpp
@@ -2,13 +2,14 @@
 #define __CMP_GRAVENUMMULTIPLIER_HPP__
 
 #include <Components/Persistent/BasePersistent.hpp>
-namespace ProceduralMaze::Cmp::Persistent {
+namespace ProceduralMaze::Cmp::Persistent
+{
 
-class GraveNumMultiplier : public BasePersistent<unsigned int>
+class GraveNumMultiplier : public BasePersistent<uint8_t>
 {
 public:
-  GraveNumMultiplier( unsigned int value = 30, unsigned int min_value = 20, unsigned int max_value = 100 )
-      : BasePersistent<unsigned int>( value, min_value, max_value )
+  GraveNumMultiplier( uint8_t value = 30, uint8_t min_value = 20, uint8_t max_value = 100 )
+      : BasePersistent<uint8_t>( value, min_value, max_value )
   {
   }
   virtual std::string class_name() const override { return "GraveNumMultiplier"; }
