+++ src/Components/Persistent/ShrineCostPerSprite.hpp
@@ -2,15 +2,16 @@
 #define __CMP_SHRINECOST_HPP__
 
 #include <Components/Persistent/BasePersistent.hpp>
-namespace ProceduralMaze::Cmp::Persistent {
+namespace ProceduralMaze::Cmp::Persistent
+{
 
 //! @brief Shrine consists of 4 sprites, so this is cost per sprite.
 //!
-class ShrineCostPerSprite : public BasePersistent<unsigned int>
+class ShrineCostPerSprite : public BasePersistent<uint8_t>
 {
 public:
-  ShrineCostPerSprite( unsigned int value = 1, unsigned int min_value = 0, unsigned int max_value = 4 )
-      : BasePersistent<unsigned int>( value, min_value, max_value )
+  ShrineCostPerSprite( uint8_t value = 1, uint8_t min_value = 0, uint8_t max_value = 4 )
+      : BasePersistent<uint8_t>( value, min_value, max_value )
   {
   }
   virtual std::string class_name() const override { return "ShrineCostPerSprite"; }
